+++ src/main/java/org/openlmis/stockmanagement/repository/SourceDestinationAssignmentRepository.java
@@ -18,15 +18,13 @@ package org.openlmis.stockmanagement.repository;
 import java.util.List;
 import java.util.UUID;
 import org.openlmis.stockmanagement.domain.sourcedestination.SourceDestinationAssignment;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.NoRepositoryBean;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 
 @NoRepositoryBean
 public interface SourceDestinationAssignmentRepository<T extends SourceDestinationAssignment>
-    extends PagingAndSortingRepository<T, UUID> {
-
-  List<T> findAll();
+    extends JpaRepository<T, UUID> {
 
   List<T> findByProgramIdAndFacilityTypeId(
       @Param("programId") UUID programId, @Param("facilityTypeId") UUID facilityTypeId);