+++ src/main/java/org/openlmis/stockmanagement/service/SourceDestinationBaseService.java
@@ -40,6 +40,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
+@SuppressWarnings("PMD.TooManyMethods")
 public abstract class SourceDestinationBaseService {
 
   @Autowired