+++ src/main/java/org/openlmis/stockmanagement/web/ValidSourceDestinationController.java
@@ -34,15 +34,17 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
+import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 
-@Controller
-@RequestMapping("/api/validDestinations")
+@RestController
+@RequestMapping("/api")
 public class ValidSourceDestinationController {
 
   private static final Logger LOGGER =
@@ -60,17 +62,17 @@ public class ValidSourceDestinationController {
   /**
    * Get a list of valid destinations.
    *
-   * @param params filtering parameters.
+   * @param parameters filtering parameters.
    * @return found valid destinations
    */
   @GetMapping(value = "/validDestinations")
-  public ResponseEntity<List<ValidSourceDestinationDto>> getValidDestinations(
-      @RequestParam ValidSourceDestinationSearchParams params) {
+  public List<ValidSourceDestinationDto> getValidDestinations(
+      @RequestParam MultiValueMap<String, String> parameters) {
+    ValidSourceDestinationSearchParams params = new ValidSourceDestinationSearchParams(parameters);
+
     LOGGER.debug(format("Try to find valid destinations with program %s and facility %s",
-        params.getProgramId().toString(), params.getFacilityId().toString()));
-    return new ResponseEntity<>(
-        validDestinationService.findDestinations(
-            params.getProgramId(), params.getFacilityId()), OK);
+        params.getProgramId(), params.getFacilityId()));
+    return validDestinationService.findDestinations(params.getProgramId(), params.getFacilityId());
   }
 
   /**
@@ -98,16 +100,18 @@ public class ValidSourceDestinationController {
   /**
    * Get a list of valid sources.
    *
-   * @param params filtering parameters.
+   * @param parameters filtering parameters.
    * @return found valid destinations
    */
   @GetMapping(value = "/validSources")
-  public ResponseEntity<List<ValidSourceDestinationDto>> getValidSources(
-      @RequestParam ValidSourceDestinationSearchParams params) {
+  public List<ValidSourceDestinationDto> getValidSources(
+      @RequestParam MultiValueMap<String, String> parameters) {
+    ValidSourceDestinationSearchParams params = new ValidSourceDestinationSearchParams(parameters);
+
     LOGGER.debug(format("Try to find valid sources with program %s and facility %s",
-        params.getProgramId().toString(), params.getFacilityId().toString()));
-    return new ResponseEntity<>(
-        validSourceService.findSources(params.getProgramId(), params.getFacilityId()), OK);
+        params.getProgramId(), params.getFacilityId()));
+    return validSourceService.findSources(
+        params.getProgramId(), params.getFacilityId());
   }
 
   /**
@@ -135,27 +139,25 @@ public class ValidSourceDestinationController {
    * Remove a valid source assignment of a program and facility type combination.
    *
    * @param assignmentId source assignment ID
-   * @return no content status
    */
   @RequestMapping(value = "/validSources/{id}", method = DELETE)
-  public ResponseEntity removeValidSourceAssignment(@PathVariable("id") UUID assignmentId) {
+  @ResponseStatus(NO_CONTENT)
+  public void removeValidSourceAssignment(@PathVariable("id") UUID assignmentId) {
     LOGGER.debug(format("Try to remove source assignment %s.", assignmentId));
     permissionService.canManageStockSources();
     validSourceService.deleteSourceAssignmentById(assignmentId);
-    return new ResponseEntity(null, NO_CONTENT);
   }
 
   /**
    * Remove a valid destination assignment of a program and facility type combination.
    *
    * @param assignmentId destination assignment ID
-   * @return no content status
    */
   @RequestMapping(value = "/validDestinations/{id}", method = DELETE)
-  public ResponseEntity removeValidDestinationAssignment(@PathVariable("id") UUID assignmentId) {
+  @ResponseStatus(NO_CONTENT)
+  public void removeValidDestinationAssignment(@PathVariable("id") UUID assignmentId) {
     LOGGER.debug(format("Try to remove destination assignment %s.", assignmentId));
     permissionService.canManageStockDestinations();
     validDestinationService.deleteDestinationAssignmentById(assignmentId);
-    return new ResponseEntity(null, NO_CONTENT);
   }
 }